# server/doc/api.yaml
openapi: 3.0.3
info:
  title: Node.js GMP Express application
  version: 1.0.0
...
paths:
...
  /api/products:
    get:
      tags:
        - product
      summary: Returns a list of products
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Returns a list of all products available
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductsResponse'
              example:
                data:
                  - id: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
                    title: Book
                    description: Interesting book
                    price: 200
                  - id: afdd68c4-d359-45e6-b9fd-c8fdb2a162a0
                    title: Pen
                    description: Cute pen
                    price: 20
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
  /api/products/{productId}:
    get:
      tags:
        - product
      summary: Returns single product
      parameters:
        - name: productId
          in: path
          description: Id (uuid) of product to return
          required: true
          schema:
            type: number
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
              example:
                data:
                  id: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
                  title: Book
                  description: Interesting book
                  price: 200
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: No product with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
components:
  schemas:
    Cart:
      type: object
      properties:
        id:
          type: string
          description: Cart id (uuid)
          example: dd5ec5ab-deaa-419c-8a6b-7e67b1f7ec87
        items:
          type: array
          description: Items added to cart
          items:
            $ref: '#/components/schemas/CartItem'
    CartItem:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        count:
          type: number
          description: Total count of specific products
          example: 2
    Product:
      type: object
      properties:
        id:
          type: string
          description: Product id (uuid)
          example: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
        title:
          type: string
          description: Product name
          example: Book
        description:
          type: string
          description: Product description
          example: Interesting book
        price:
          type: number
          description: Product price
          example: 200
...
    CartResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            cart:
              $ref: '#/components/schemas/Cart'
            totalPrice:
              type: number
        error:
          $ref: '#/components/schemas/ErrorResponse'
...
    ProductResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Product'
        error:
          $ref: '#/components/schemas/ErrorResponse'
    ProductsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        error:
          $ref: '#/components/schemas/ErrorResponse'
...
    ErrorResponse:
      type: object
      nullable: true
      properties:
        message:
          type: string